debug checks
git-svn-id: http://localhost/svn/entropybroker/trunk@555 6e73ccbf-f98f-4734-8601-04fa8acba097

+++ encrypt_stream.h
@@ -14,6 +14,6 @@ public:
 
 	virtual std::string get_name() = 0;
 
-	virtual void encrypt(unsigned char *p_in, size_t len, unsigned char *p_out) = 0;
-	virtual void decrypt(unsigned char *p_in, size_t len, unsigned char *p_out) = 0;
+	virtual void encrypt(unsigned char *p_in, int len, unsigned char *p_out) = 0;
+	virtual void decrypt(unsigned char *p_in, int len, unsigned char *p_out) = 0;
 };
